feat(portal-base): add --only flag to yarn start
